requirements.py3.txt
# PepperChat Python2 Requirements
python-dotenv
zmq
openai
websocket-client
